Count matrix

An S4 class to represent a count matrix.

CountMatrix(data = 0, nrow = 1, ncol = 1, byrow = FALSE,
  dimnames = NULL, ...)

Arguments

data
an optional data vector (including a list or expression vector). Non-atomic classed R objects are coerced by as.vector and all attributes discarded.

nrow
the desired number of rows.

ncol
the desired number of columns.

byrow
logical. If FALSE (the default) the matrix is filled by columns, otherwise the matrix is filled by rows.

dimnames
A dimnames attribute for the matrix: NULL or a list of length 2 giving the row and column names respectively. An empty list is treated as NULL, and a list of length one as row names. The list can be named, and the list names will be used as names for the dimensions.

...
additional arguments to be passed to or from methods.

Value

TODO

Note

Numeric values are rounded to zero decimal places and then coerced to integer as by as.integer.

Matrix ID

When a matrix is first created, an identifier is generated (UUID v4). This ID is preserved when coercing to another class. Thus, the object ID is unique within the same class, but two objects of different classes can have the same ID. This makes it possible to identify objects representing the same initial data and associate them with the results of specific computations.

Get and set

In the code snippets below, x is a *Matrix object.

get_id(x)

Get the ID of x.

Access

In the code snippets below, x is a *Matrix object.

dim(x)

Returns the dimension of x.

nrow(x)

Returns the number of rows present in x.

ncol(x)

Returns the number of columns present in x.

dimnames(x), dimnames(x) <- value

Retrieves or sets the row dimnames of x according to value.

rownames(x), rownames(x) <- value

Retrieves or sets the row names of x according to value.

colnames(x), colnames(x) <- value

Retrieves or sets the column names of x according to value.

Subset

In the code snippets below, x is a *Matrix object.

x[i, j]

Extracts elements selected by subscripts i and j. Indices are numeric, integer or character vectors or empty (missing) or NULL. Numeric values are coerced to integer as by as.integer (and hence truncated towards zero). Character vectors will be matched to the name of the elements. An empty index (a comma separated blank) indicates that all entries in that dimension are selected. Returns an object of the same class as x.

x[[i]]

Extracts informations from a slot selected by subscript i. i should be one of "id" or NULL.

See also

Examples

## Create a count data matrix
A1 <- CountMatrix(data = sample(0:10, 100, TRUE),
                  nrow = 10, ncol = 10, byrow = TRUE)

## Access
get_id(A1)

#> [1] "064862bf-2c8a-4646-bb97-8c1d00f287c4"

dim(A1) # Get the matrix dimensions

#> [1] 10 10

colnames(A1) # Get the column names

#>  [1] "V1"  "V2"  "V3"  "V4"  "V5"  "V6"  "V7"  "V8"  "V9"  "V10"

colnames(A1) <- letters[11:20] # Set the column names
rownames(A1) # Get the rownames

#>  [1] "1"  "2"  "3"  "4"  "5"  "6"  "7"  "8"  "9"  "10"

rownames(A1) <- LETTERS[1:10] # Set the row names

## Subset
A1[["id"]] # Get the matrix ID

#> [1] "064862bf-2c8a-4646-bb97-8c1d00f287c4"

A1[, ] # Get all values

#>   k  l  m n o p  q  r  s t
#> A 6  4  5 3 5 8  4  4  7 1
#> B 6 10 10 4 4 1  5  3  1 2
#> C 8  5 10 6 4 5  8  5  9 4
#> D 1  0  3 7 5 0  5  1  2 0
#> E 8  1  4 7 0 2 10  9  2 1
#> F 6  9  2 5 3 9  0  6  5 5
#> G 2  8  1 7 0 0  6  5  8 9
#> H 1  5  6 9 5 4  6  2  0 4
#> I 0  5  0 0 1 4  6 10  0 3
#> J 6  6  7 1 7 8  2  8 10 5

A1[1, ] # Get the first row

#> k l m n o p q r s t 
#> 6 4 5 3 5 8 4 4 7 1 

A1[c("A", "B", "C"), ] # Get the first three rows

#>   k  l  m n o p q r s t
#> A 6  4  5 3 5 8 4 4 7 1
#> B 6 10 10 4 4 1 5 3 1 2
#> C 8  5 10 6 4 5 8 5 9 4

A1[c("A", "B", "C"), 1] # Get the first three rows of the first column

#> A B C 
#> 6 6 8 

A1[, 1, drop = FALSE] # Get the first column

#>   k
#> A 6
#> B 6
#> C 8
#> D 1
#> E 8
#> F 6
#> G 2
#> H 1
#> I 0
#> J 6

## Coerce counts to frequencies
B <- as_frequency(A1)
## Row sums are internally stored before coercing to a frequency matrix
get_totals(B) # Get row sums

#>  A  B  C  D  E  F  G  H  I  J 
#> 47 46 64 24 44 50 46 42 29 60 

## This allows to restore the source data
A2 <- as_count(B)
all(A1 == A2)

#> [1] TRUE