Space and Time

An S4 class to represent space-time informations.

Slots

dates

A two column numeric matrix giving the date value and error, respectively.

coordinates

A three columns numeric matrix (x, y and z) giving the geographic coordinates (longitude, latitude and elevation, respectively).

epsg

An integer giving the EPSG code of the spatial reference system used. Numeric values are coerced to integer as by as.integer (and hence truncated towards zero).

Get and set

In the code snippets below, x is a SpaceTime object.

get_dates(x), set_dates(x) <- value

Get or set the dates of x according to value (see set_dates for details).

get_coordinates(x), set_coordinates(x) <- value

Get or set the geographical coordinates of x according to value (see set_coordinates for details).

get_epsg(x), set_epsg(x) <- value

Get or set the EPSG of x according to value. Numeric values are coerced to integer as by as.integer (and hence truncated towards zero).

as_features(x)

Convert an AbundanceMatrix object to a data frame with dates and coordinates columns.