Extract or Replace Parts of an Object

Operators acting on objects to extract or replace parts.

# S4 method for SpaceTime
[[(x, i)

# S4 method for CountMatrix
[[(x, i)

# S4 method for FrequencyMatrix
[[(x, i)

# S4 method for IncidenceMatrix
[[(x, i)

# S4 method for OccurrenceMatrix
[[(x, i)

# S4 method for SimilarityMatrix
[[(x, i)

Arguments

x
An object from which to extract element(s) or in which to replace element(s).

i
Indices specifying elements to extract or replace. Indices are numeric, integer or character vectors or empty (missing) or NULL. Numeric values are coerced to integer as by as.integer (and hence truncated towards zero). Character vectors will be matched to the name of the elements. An empty index (a comma separated blank) indicates that all entries in that dimension are selected.

Value

A subsetted object.

See also

Examples

## Create a count data matrix
A1 <- CountMatrix(data = sample(0:10, 100, TRUE),
                  nrow = 10, ncol = 10, byrow = TRUE)

## Access
get_id(A1)

#> [1] "0298d8b2-fa10-44e0-9f82-804ac82679d9"

dim(A1) # Get the matrix dimensions

#> [1] 10 10

colnames(A1) # Get the column names

#>  [1] "V1"  "V2"  "V3"  "V4"  "V5"  "V6"  "V7"  "V8"  "V9"  "V10"

colnames(A1) <- letters[11:20] # Set the column names
rownames(A1) # Get the rownames

#>  [1] "1"  "2"  "3"  "4"  "5"  "6"  "7"  "8"  "9"  "10"

rownames(A1) <- LETTERS[1:10] # Set the row names

## Subset
A1[["id"]] # Get the matrix ID

#> [1] "0298d8b2-fa10-44e0-9f82-804ac82679d9"

A1[, ] # Get all values

#>    k  l  m n o  p  q r  s  t
#> A  2  4  3 7 2  6  6 7  7  3
#> B  8  5  1 5 9 10  7 4  0  1
#> C  0  2  5 8 1  5  7 4  2  9
#> D 10  8  6 4 7  9  3 6 10  3
#> E  9  5  0 5 5 10  1 5  7 10
#> F  5  2  6 7 6  4  5 3  3  8
#> G  3 10 10 8 8  7 10 3  7  1
#> H  6  4  3 6 3  6  0 0  7 10
#> I  2  9  8 2 6  6  4 5  7  6
#> J  5  9  7 1 1  4  4 9  2  4

A1[1, ] # Get the first row

#> k l m n o p q r s t 
#> 2 4 3 7 2 6 6 7 7 3 

A1[c("A", "B", "C"), ] # Get the first three rows

#>   k l m n o  p q r s t
#> A 2 4 3 7 2  6 6 7 7 3
#> B 8 5 1 5 9 10 7 4 0 1
#> C 0 2 5 8 1  5 7 4 2 9

A1[c("A", "B", "C"), 1] # Get the first three rows of the first column

#> A B C 
#> 2 8 0 

A1[, 1, drop = FALSE] # Get the first column

#>    k
#> A  2
#> B  8
#> C  0
#> D 10
#> E  9
#> F  5
#> G  3
#> H  6
#> I  2
#> J  5

## Coerce counts to frequencies
B <- as_frequency(A1)
## Row sums are internally stored before coercing to a frequency matrix
get_totals(B) # Get row sums

#>  A  B  C  D  E  F  G  H  I  J 
#> 47 50 43 66 57 49 67 45 55 46 

## This allows to restore the source data
A2 <- as_count(B)
all(A1 == A2)

#> [1] TRUE